Practice ROBOT/requirement.txt
Appium-Python-Client==4.0.1
attrs==23.2.0
certifi==2024.7.4
cffi==1.16.0
click==8.1.7
colorama==0.4.6
decorator==5.1.1
docutils==0.21.2
h11==0.14.0
idna==3.7
kitchen==1.2.6
numpy==2.0.1
opencv-python==4.10.0.84
outcome==1.3.0.post0
pillow==10.4.0
pycparser==2.22
PySocks==1.7.1
robotbackgroundlogger==1.2
robotframework==7.0.1
robotframework-appiumlibrary==2.1.0
robotframework-pythonlibcore==4.4.1
robotframework-seleniumlibrary==6.5.0
selenium==4.23.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0